Fix: LT-18429 Fix 32-bit window handle in VwSearchKiller
* Code changed in ViewsTlb.idl to generate long Window in Views.cs

+++ Src/Generic/UtilString.cpp
@@ -1639,7 +1639,7 @@ template<typename XChar>
 	AssertArrayN(prgchIns, cchIns);
 
 	int cchCur = m_pbuf->Cch();
-	Assert((uint)ichMin <= (uint)ichLim && (uint)ichLim <= (uint)cchCur);
+	Assert(ichMin <= ichLim && ichLim <= cchCur);
 
 	if (!cchIns)
 	{
